Fixed file locations after updating inwave functionality
Accidentally put some files in the wrong places when updating the inwave functionality causing duplicate files to be created, so just moved files around and placed them where they needed to be.
